Update the README.md for external access
The instructions for accessing the redis cluster from outside of the
cloud platform VPC were out of date. All elasticache instances are
encrypted as a matter of policy, and the instructions as written would
not work.

This change uses the same procedure as that in the user guide for
accessing RDS instances, and also gives ruby code examples, because the
`redis-cli` does not work with encrypted redis connnections.